+++ src/lib/gtfs-ride/board-alights.ts
@@ -1,9 +1,9 @@
-import {
+import type {
   QueryOptions,
   SqlOrderBy,
-  SqlResults,
-  SqlSelect,
+  QueryResult,
   SqlWhere,
+  BoardAlight,
 } from '../../types/global_interfaces.ts';
 import { openDb } from '../db.ts';
 import {
@@ -15,12 +15,12 @@ import {
 /*
  * Returns an array of all board-alights that match the query parameters.
  */
-export function getBoardAlights(
+export function getBoardAlights<Fields extends keyof BoardAlight>(
   query: SqlWhere = {},
-  fields: SqlSelect = [],
+  fields: Fields[] = [],
   orderBy: SqlOrderBy = [],
   options: QueryOptions = {},
-): SqlResults {
+) {
   const db = options.db ?? openDb();
   const tableName = 'board_alight';
   const selectClause = formatSelectClause(fields);
@@ -31,5 +31,5 @@ export function getBoardAlights(
     .prepare(
       `${selectClause} FROM ${tableName} ${whereClause} ${orderByClause};`,
     )
-    .all() as SqlResults;
+    .all() as QueryResult<BoardAlight, Fields>[];
 }