+++ src/lib/gtfs-ride/ride-feed-info.ts
@@ -1,9 +1,9 @@
-import {
+import type {
   QueryOptions,
   SqlOrderBy,
-  SqlResults,
-  SqlSelect,
+  QueryResult,
   SqlWhere,
+  RideFeedInfo,
 } from '../../types/global_interfaces.ts';
 import { openDb } from '../db.ts';
 import {
@@ -15,12 +15,12 @@ import {
 /*
  * Returns an array of all ride-feed-info that match the query parameters.
  */
-export function getRideFeedInfo(
+export function getRideFeedInfo<Fields extends keyof RideFeedInfo>(
   query: SqlWhere = {},
-  fields: SqlSelect = [],
+  fields: Fields[] = [],
   orderBy: SqlOrderBy = [],
   options: QueryOptions = {},
-): SqlResults {
+) {
   const db = options.db ?? openDb();
   const tableName = 'ride_feed_info';
   const selectClause = formatSelectClause(fields);
@@ -31,5 +31,5 @@ export function getRideFeedInfo(
     .prepare(
       `${selectClause} FROM ${tableName} ${whereClause} ${orderByClause};`,
     )
-    .all() as SqlResults;
+    .all() as QueryResult<RideFeedInfo, Fields>[];
 }