+++ src/lib/gtfs-ride/rider-trips.ts
@@ -1,9 +1,9 @@
-import {
+import type {
   QueryOptions,
   SqlOrderBy,
-  SqlResults,
-  SqlSelect,
+  QueryResult,
   SqlWhere,
+  RiderTrip,
 } from '../../types/global_interfaces.ts';
 import { openDb } from '../db.ts';
 import {
@@ -15,12 +15,12 @@ import {
 /*
  * Returns an array of all rider trips that match the query parameters.
  */
-export function getRiderTrips(
+export function getRiderTrips<Fields extends keyof RiderTrip>(
   query: SqlWhere = {},
-  fields: SqlSelect = [],
+  fields: Fields[] = [],
   orderBy: SqlOrderBy = [],
   options: QueryOptions = {},
-): SqlResults {
+) {
   const db = options.db ?? openDb();
   const tableName = 'rider_trip';
   const selectClause = formatSelectClause(fields);
@@ -31,5 +31,5 @@ export function getRiderTrips(
     .prepare(
       `${selectClause} FROM ${tableName} ${whereClause} ${orderByClause};`,
     )
-    .all() as SqlResults;
+    .all() as QueryResult<RiderTrip, Fields>[];
 }