+++ src/lib/gtfs-ride/trip-capacities.ts
@@ -1,9 +1,9 @@
-import {
+import type {
   QueryOptions,
   SqlOrderBy,
-  SqlResults,
-  SqlSelect,
+  QueryResult,
   SqlWhere,
+  TripCapacity,
 } from '../../types/global_interfaces.ts';
 import { openDb } from '../db.ts';
 import {
@@ -15,12 +15,12 @@ import {
 /*
  * Returns an array of all trip-capacities that match the query parameters.
  */
-export function getTripCapacities(
+export function getTripCapacities<Fields extends keyof TripCapacity>(
   query: SqlWhere = {},
-  fields: SqlSelect = [],
+  fields: Fields[] = [],
   orderBy: SqlOrderBy = [],
   options: QueryOptions = {},
-): SqlResults {
+) {
   const db = options.db ?? openDb();
   const tableName = 'trip_capacity';
   const selectClause = formatSelectClause(fields);
@@ -31,5 +31,5 @@ export function getTripCapacities(
     .prepare(
       `${selectClause} FROM ${tableName} ${whereClause} ${orderByClause};`,
     )
-    .all() as SqlResults;
+    .all() as QueryResult<TripCapacity, Fields>[];
 }