+++ src/lib/non-standard/timetable-pages.ts
@@ -1,9 +1,9 @@
-import {
+import type {
   QueryOptions,
   SqlOrderBy,
-  SqlResults,
-  SqlSelect,
+  QueryResult,
   SqlWhere,
+  TimetablePage,
 } from '../../types/global_interfaces.ts';
 import { openDb } from '../db.ts';
 import {
@@ -15,12 +15,12 @@ import {
 /*
  * Returns an array of all timetable pages that match the query parameters.
  */
-export function getTimetablePages(
+export function getTimetablePages<Fields extends keyof TimetablePage>(
   query: SqlWhere = {},
-  fields: SqlSelect = [],
+  fields: Fields[] = [],
   orderBy: SqlOrderBy = [],
   options: QueryOptions = {},
-): SqlResults {
+) {
   const db = options.db ?? openDb();
   const tableName = 'timetable_pages';
   const selectClause = formatSelectClause(fields);
@@ -31,5 +31,5 @@ export function getTimetablePages(
     .prepare(
       `${selectClause} FROM ${tableName} ${whereClause} ${orderByClause};`,
     )
-    .all() as SqlResults;
+    .all() as QueryResult<TimetablePage, Fields>[];
 }