+++ src/types/global_interfaces.ts
@@ -74,8 +74,6 @@ export type SqlValue =
 
 export type SqlWhere = Record<string, null | SqlValue | SqlValue[]>;
 
-export type SqlSelect = string[];
-
 export type QueryResult<Base extends object, Select extends keyof Base> = [
   Select,
 ] extends [never]
@@ -88,8 +86,6 @@ export interface QueryOptions {
   db?: Database;
   bounding_box_side_m?: number;
 }
-
-export type SqlResults = Array<Record<string, any>>;
 
 export interface Agency {
   agency_id?: string;
@@ -390,4 +386,301 @@ export interface Trip {
   shape_id?: string;
   wheelchair_accessible?: 0 | 1 | 2;
   bikes_allowed?: 0 | 1 | 2;
+}
+
+export interface Timetable {
+  timetable_id?: string;
+  route_id?: string;
+  direction_id?: 0 | 1;
+  start_date?: number;
+  end_date?: number;
+  monday?: 0 | 1;
+  tuesday?: 0 | 1;
+  wednesday?: 0 | 1;
+  thursday?: 0 | 1;
+  friday?: 0 | 1;
+  saturday?: 0 | 1;
+  sunday?: 0 | 1;
+  start_time?: string;
+  start_timestamp?: number;
+  end_time?: string;
+  end_timestamp?: number;
+  timetable_label?: string;
+  service_notes?: string;
+  orientation?: string;
+  timetable_page_id?: string;
+  timetable_sequence?: number;
+  direction_name?: string;
+  include_exceptions?: 0 | 1;
+  show_trip_continuation?: 0 | 1;
+}
+
+export interface TimetableStopOrder {
+  timetable_id: string;
+  stop_id: string;
+  stop_sequence: number;
+}
+
+export interface TimetableNote {
+  note_id: string;
+  symbol: string;
+  note: string;
+}
+
+export interface TimetableNotesReference {
+  timetable_id: string;
+  route_id?: string;
+  trip_id?: string;
+  stop_id?: string;
+  stop_sequence?: number;
+  show_on_stoptime?: 0 | 1;
+}
+
+export interface TimetablePage {
+  timetable_page_id: string;
+  timetable_page_label: string;
+  filename: string;
+}
+
+export interface TripsDatedVehicleJourney {
+  trip_id: string;
+  operating_day_date: string;
+  dated_vehicle_journey_gid: string;
+  journey_number: number;
+}
+
+export interface DeadheadTime {
+  deadhead_id: string;
+  arrival_time: string;
+  arrival_timestamp: number;
+  departure_time: string;
+  departure_timestamp: number;
+  ops_location_id?: string;
+  stop_id?: string;
+  location_sequence: number;
+  shape_dist_traveled?: number;
+}
+
+export interface Deadhead {
+  deadhead_id: string;
+  service_id: string;
+  block_id: string;
+  shape_id?: string;
+  to_trip_id?: string;
+  from_trip_id?: string;
+  to_deadhead_id?: string;
+  from_deadhead_id?: string;
+}
+
+export interface OpsLocation {
+  ops_location_id: string;
+  ops_location_code?: string;
+  ops_location_name: string;
+  ops_location_desc?: string;
+  ops_location_lat: number;
+  ops_location_lon: number;
+}
+
+export interface RunEvent {
+  run_event_id: string;
+  piece_id: string;
+  event_type: number;
+  event_name?: string;
+  event_time: string;
+  event_duration: number;
+  event_from_location_type: 0 | 1;
+  event_from_location_id?: string;
+  event_to_location_type: 0 | 1;
+  event_to_location_id?: string;
+}
+
+export interface RunPiece {
+  run_id: string;
+  piece_id: string;
+  start_type: 0 | 1 | 2;
+  start_trip_id: string;
+  start_trip_position?: number;
+  end_type: 0 | 1 | 2;
+  end_trip_id: string;
+  end_trip_position?: number;
+}
+
+export interface ServiceAlert {
+  cause: number;
+  start_time: string;
+  end_time: string;
+  headline: string;
+  description: string;
+  created_timestamp: number;
+  expiration_timestamp: number;
+}
+
+export interface StopTimeUpdate {
+  trip_id?: string;
+  trip_start_time?: string;
+  direction_id?: 0 | 1;
+  route_id?: string;
+  stop_id?: string;
+  stop_sequence?: number;
+  arrival_delay?: number;
+  departure_delay?: number;
+  departure_timestamp?: string;
+  arrival_timestamp?: string;
+  schedule_relationship?: string;
+  created_timestamp: number;
+  expiration_timestamp: number;
+}
+
+export interface TripUpdate {
+  update_id: string;
+  vehicle_id?: string;
+  trip_id?: string;
+  trip_start_time?: string;
+  direction_id?: 0 | 1;
+  route_id?: string;
+  start_date?: number;
+  timestamp?: string;
+  schedule_relationship?: string;
+  created_timestamp: number;
+  expiration_timestamp: number;
+}
+
+export interface VehiclePosition {
+  update_id: string;
+  bearing?: number;
+  latitude?: number;
+  longitude?: number;
+  speed?: number;
+  current_stop_sequence?: number;
+  trip_id?: string;
+  trip_start_date?: number;
+  trip_start_time?: string;
+  congestion_level?: string;
+  occupancy_status?: string;
+  occupancy_percentage?: number;
+  vehicle_stop_status?: string;
+  vehicle_id?: string;
+  vehicle_label?: string;
+  vehicle_license_plate?: string;
+  vehicle_wheelchair_accessible?: number;
+  timestamp?: string;
+  created_timestamp: number;
+  expiration_timestamp: number;
+}
+
+export interface BoardAlight {
+  trip_id: string;
+  stop_id: string;
+  stop_sequence: number;
+  record_use: 0 | 1;
+  schedule_relationship?: number;
+  boardings?: number;
+  alightings?: number;
+  current_load?: number;
+  load_count?: number;
+  load_type?: number;
+  rack_down?: number;
+  bike_boardings?: number;
+  bike_alightings?: number;
+  ramp_used?: number;
+  ramp_boardings?: number;
+  ramp_alightings?: number;
+  service_date?: number;
+  service_arrival_time?: string;
+  service_arrival_timestamp?: number;
+  service_departure_time?: string;
+  service_departure_timestamp?: number;
+  source?: 0 | 1 | 2 | 3 | 4;
+}
+
+export interface RideFeedInfo {
+  ride_files: number;
+  ride_start_date?: number;
+  ride_end_date?: number;
+  gtfs_feed_date?: number;
+  default_currency_type?: string;
+  ride_feed_version?: string;
+}
+
+export interface RiderTrip {
+  rider_id: string;
+  agency_id?: string;
+  trip_id?: string;
+  boarding_stop_id?: string;
+  boarding_stop_sequence?: number;
+  alighting_stop_id?: string;
+  alighting_stop_sequence?: number;
+  service_date?: number;
+  boarding_time?: string;
+  boarding_timestamp?: number;
+  alighting_time?: string;
+  alighting_timestamp?: number;
+  rider_type?: number;
+  rider_type_description?: string;
+  fare_paid?: number;
+  transaction_type?: number;
+  fare_media?: number;
+  accompanying_device?: number;
+  transfer_status?: number;
+}
+
+export interface Ridership {
+  total_boardings: number;
+  total_alightings: number;
+  ridership_start_date?: number;
+  ridership_end_date?: number;
+  ridership_start_time?: string;
+  ridership_start_timestamp?: number;
+  ridership_end_time?: string;
+  ridership_end_timestamp?: number;
+  service_id?: string;
+  monday?: 0 | 1;
+  tuesday?: 0 | 1;
+  wednesday?: 0 | 1;
+  thursday?: 0 | 1;
+  friday?: 0 | 1;
+  saturday?: 0 | 1;
+  sunday?: 0 | 1;
+  agency_id?: string;
+  route_id?: string;
+  direction_id?: 0 | 1;
+  trip_id?: string;
+  stop_id?: string;
+}
+
+export interface TripCapacity {
+  agency_id?: string;
+  trip_id?: string;
+  service_date?: number;
+  vehicle_description?: string;
+  seated_capacity?: number;
+  standing_capacity?: number;
+  wheelchair_capacity?: number;
+  bike_capacity?: number;
+}
+
+export interface CalendarAttribute {
+  service_id: string;
+  service_description: string;
+}
+
+export interface Direction {
+  route_id: string;
+  direction_id?: 0 | 1;
+  direction: string;
+}
+
+export interface RouteAttribute {
+  route_id: string;
+  category: number;
+  subcategory: number;
+  running_way: number;
+}
+
+export interface StopAttribute {
+  stop_id: string;
+  accessibility_id?: number;
+  cardinal_direction?: string;
+  relative_position?: string;
+  stop_city?: string;
 }